% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ei.R
\name{checkZeligEIna.action}
\alias{checkZeligEIna.action}
\title{Checking function for argument to na.action}
\usage{
checkZeligEIna.action(na.action)
}
\description{
Checking function for argument to na.action
}
\keyword{internal}
