% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ei.R
\name{convertEIformula2}
\alias{convertEIformula2}
\title{Conversion utility to allow different possible formula notations, and deal with zeroes and missing values, for EI models in eiml, eirxc}
\usage{
convertEIformula2(formula, data, N, na.action, rxc = FALSE)
}
\description{
Conversion utility to allow different possible formula notations, and deal with zeroes and missing values, for EI models in eiml, eirxc
}
\keyword{internal}
