% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dynamic-ei.R
\docType{class}
\name{Zelig-eidynamic-class}
\alias{Zelig-eidynamic-class}
\alias{zeidynamic}
\title{Quinn's Dynamic Ecological Inference Model}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zeligei-eidynamic.html}
}

