% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-hier-ei.R
\docType{class}
\name{Zelig-eihier-class}
\alias{Zelig-eihier-class}
\alias{zeihier}
\title{Wakefield's Hierarchical Ecological Inference Model}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zeligei-eihier.html}
}

