% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ml-ei.R
\docType{class}
\name{Zelig-eiml-class}
\alias{Zelig-eiml-class}
\alias{zeiml}
\title{Ecological Inference Model by Maximum Likelihood}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zeligei_eiml.html}
}

