% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-rxc-ei.R
\docType{class}
\name{Zelig-eirxc-class}
\alias{Zelig-eirxc-class}
\alias{zeirxc}
\title{Multinomial Dirichlet model for Ecological Inference in RxC tables}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zeligei_eirxc.html}
}
\section{Methods}{

\describe{
\item{\code{getcoef()}}{Get estimated model coefficients}
}}

