% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ei.R
\name{convertEIformula}
\alias{convertEIformula}
\title{Conversion utility to allow different possible formula notations for EI models}
\usage{
convertEIformula(formula, N, data)
}
\description{
Conversion utility to allow different possible formula notations for EI models
}
\keyword{internal}

