% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ei.R
\docType{class}
\name{Zelig-ei-class}
\alias{Zelig-ei-class}
\alias{zei}
\title{Ecological Inference object for inheritance across models in ZeligEI}
\description{
Ecological Inference object for inheritance across models in ZeligEI
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig command estimates a variety of statistical models}
}}

